requirements.txt
Flask
Flask-SQLAlchemy
Flask-Mail
python-dotenv
cloudinary
reportlab
psycopg2-binary
gunicorn
Flask
pyparsing